Passing part 2
This involves detecting cycles (in the manner of GOL flyers) and
fast-forwarding to the required number of steps.

I suppose it is possible for there not to be a cycle, but the input we
are given has a one after only a hundred or so cycles and it is a simple
flyer of order-1.